refactor(vnpy_bridge): clean up __init__, tighten _core type
- Use **kwargs instead of listing vnpy params (pure pass-through)
- self._core: Strategy[Any] | None (was Any | None)
- Remove redundant self._strategy_name (use vnpy's self.strategy_name)
- Remove redundant self.price_tick (vnpy CtaTemplate handles it)

refactor(core): move fills from Strategy to Bridge, UninitializedStrategy sentinel

Strategy no longer manages fills — Bridge is the canonical source (industry
standard). Strategy.on_fill() only updates position/entry_price state.

- core/base.py: UninitializedStrategy sentinel replaces None for _core
  All abstract methods raise RuntimeError, fills returns []
- ma_strategy.py: remove _fills, fills property, on_fill no longer appends
- vnpy_bridge / tqsdk_bridge: self._fills list, append on each fill
- cli/commands/test.py: local fills list instead of strategy.fills

refactor: fills now managed by Bridge, Strategy only updates state

- Remove _fills/fills from MaStrategyCore (ABC UninitializedStrategy too)
- vnpy_bridge/tqsdk_bridge: self._fills list, public bridge.fills property
- backtest.py: bridge.fills instead of getattr(strategy_core, 'fills', [])
- test_strategies.py: remove fill-accumulation tests (behavior moved to Bridge)
- base.py: update doc, remove Strategy.fills default property

fix(backtest): BacktestFinished import, indentation, missing args

- except BacktestFinished -> except tqsdk.BacktestFinished (import removed)
- Fix indentation of inner except clause (0 -> 12 spaces, was syntax error)
- Add missing statistics/start_date/end_date to failed-branch insert_backtest

feat(common): BacktestResult dataclass replaces dict[str, object] pipeline

- common/types.py: BacktestResult dataclass with typed fields for all
  backtest results (performance, engine config, fills, daily data)
- data/store.py: insert_backtest_detailed accepts BacktestResult,
  removes fuzzy dict key lookups + stale import
- data/manager.py: insert_backtest accepts BacktestResult (10+ params -> 1)
- backtest/vnpy_backtest_engine.py: run() returns list[BacktestResult]
  instead of list[dict], maps vnpy statistics to typed fields
- backtest/optimizer.py: use r.sharpe_ratio/r.success instead of
  r['statistics'].get('sharpe_ratio') / r.get('success')
- cli/commands/backtest.py: all insert_backtest callers construct
  BacktestResult; _persist_search_results mutates fields in-place

fix: BacktestResult dict→dataclass 遗留问题 + 数值字段 null 归一化

- backtest/vnpy_backtest_engine.py: r['success'] → r.success（BacktestResult
  dataclass 不支持下标访问），类型注解 list[dict] → list[BacktestResult]
- data/store.py: insert_backtest_trades 的 quantity 类型 int→float 同步
- data/store.py: get_backtests_for_run() 对 total_return/sharpe_ratio/
  max_drawdown/win_rate/total_trades 做 float(x or 0) 归一化，避免
  前端 .toFixed() 在 null 上崩溃
- data/models.py: TradeRecord.quantity / BacktestTrade.quantity int→float
- strategies/core/types.py: Signal/StrategyPosition/Fill 的 volume int→float

+++ data/models.py
@@ -69,14 +69,14 @@ class TradeRecord(BaseModel):
     offset: str = "open"
     open_price: float
     close_price: float
-    quantity: int
+    quantity: float
     pnl: float = 0.0
     commission: float = 0.0
     created_at: str | None = None
 
     @field_validator('quantity')
     @classmethod
-    def validate_quantity(cls, v: int) -> int:
+    def validate_quantity(cls, v: float) -> float:
         if v <= 0:
             raise ValueError('quantity must be greater than 0')
         return v
@@ -245,7 +245,7 @@ class BacktestTrade(OrmBaseModel):
     offset: CharField = CharField()
     open_price: FloatField = FloatField()
     close_price: FloatField = FloatField()
-    quantity: IntegerField = IntegerField()
+    quantity: FloatField = FloatField()
     pnl: FloatField = FloatField()
     commission: FloatField = FloatField()
     created_at: DateTimeField = DateTimeField(constraints=[SQL('DEFAULT CURRENT_TIMESTAMP')])
